+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Updated to use `@shopgate/cloud-sdk-webpack@1.20.12` that brings support for Babel 7.
 * Updated vulnerable dependencies and fixed their versions
 * Increased extensions list maximum height
+* Attached extensions config is now actualized during backend start
 * Fixed extension config changes watcher
 * Fixed update notifier behaviour
 * Fix which prevents running `sgconnect` commands in user's home directory